+++ components/building/building-dialog.tsx
@@ -30,7 +30,7 @@ export function BuildingDialog() {
         </Button>
       </DialogTrigger>
 
-      <DialogContent className="flex flex-col gap-0 p-0 sm:max-h-[min(640px,80vh)] sm:max-w-[600px] [&>button:last-child]:top-3.5">
+      <DialogContent className="flex flex-col gap-0 p-0 sm:max-h-[min(640px,80vh)] sm:max-w-[600px] [&>button:last-child]:top-3.5 bg-background-100">
         <DialogHeader className="contents space-y-0 text-left">
           <DialogTitle className="border-b px-6 py-4 text-base">
             Update Workshops